AI Powered DocumentBook Intelligence System/requirements.txt

streamlit>=1.20.0
python-dotenv
transformers
sentence-transformers
langchain
langchain-community
faiss-cpu
chromadb
mistralai
langchain-mistralai
